agrego evidencia general

+++ 251dswSpringBackendTallerAutomotriz/src/main/java/sm/dswTaller/ms/tallerAutomotriz/controller/AutoController.java
@@ -90,7 +90,7 @@ public class AutoController {
        
     }
     @GetMapping("/{id}")
-    public ResponseEntity<?> getAutoById(@PathVariable Integer id) {
+    public ResponseEntity<?> getAutoById(@PathVariable Long id) {
         AutoDTO dto = autoService.getAutoById(id);
         return ResponseEntity.ok(dto);
     }
